directfb-examples: remove options for each example
For some historical reasons, the directfb-examples package has one
sub-option for each example, and the .mk files go through some great
lengths to install only the needed data files for each example.

This is a bit silly for several reasons:

 * In no other Buildroot package we do offer such a fine-grained
   selection of what the package installs;

 * directfb-examples are examples, they are therefore typically not
   meant to be used on a final production system, but more during
   development or for demonstration. Those are situations where size
   generally isn't that critical (all installed, the package takes 2.4
   MB on the rootfs);

 * Those options and the .mk file logic has to be updated whenever the
   package is bumped;

 * Users can always use a post-build script to keep only the examples
   they are interested in.

All the options are moved to Config.in.legacy. However, we could
debate on whether this is actually useful, since the main option
BR2_PACKAGE_DIRECTFB_EXAMPLES will now install a superset of whatever
was selected before through sub-options.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>