modem-manager: select libgudev, if systemd is enabled
When eudev is used as the udev provider, libgudev is automatically
provided as it is part of eudev. However, when systemd is used as
the udev provider, libgudev is not provided, and needs to be built
separately. This is why we select the libgudev package only if
systemd is used.

Fixes: http://autobuild.buildroot.net/results/d59/d597a81271a082c8252e2333906815c437b6576d/

While at it, group the existing dbus select together with the dbus-glib
select, so that all "select" statements are together.

Signed-off-by: Yegor Yefremov <yegorslists@googlemail.com>
Reviewed-by: Vicente Olivert Riera <Vincent.Riera@imgtec.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>