libevdev: bump version to 1.4.4
>From [1]:

A couple of cleanups, one notable fix: we now transparently fix invalid
ABS_MT_TRACKING_ID ranges to something sensible.

The touchpad-edge-detector tool prints out suggested hwdb rules for
resolution overrides. And since the resolution is the main thing to fix on a
device, the libevdev-tweak-device now has a --resolution argument to quickly
set the resolution on the x/y axes.

[1] http://lists.freedesktop.org/archives/input-tools/2015-August/001210.html

Signed-off-by: Peter Seiderer <ps.report@gmx.net>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>