util-linux: disable systemd support
Fixes (works around) #8421

Util-linux has optional systemd support, but we cannot enable it as systemd
depends on util-linux, so that would create a circular dependency.

The systemd unit directory location detection also fails as pkg-config
0.9.12+ prefixes all directory variables with the sysroot, and the configure
script expects to see the target location, so disable that as well for good
measure.

The systemd support doesn't seem to add much, so it isn't a big loss.

Signed-off-by: Peter Korsgaard <peter@korsgaard.com>