configs: remove calao_snowball_defconfig
This defconfig has not been updated since a long time, and it now
fails to build because the latest version of U-Boot has removed for
the Snowball because nobody converted it to the new U-Boot
standards. And our defconfig was not specifying an explicit U-Boot
version for it.

On the kernel side, the Git repository has not seen any commits since
3 years, and generally speaking, the Snowball project and the SoC it
is based on are no longer being developed.

Consequently, let's get rid of this defconfig altogether.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>