rrdtool: bump to version 1.5.5
Drop the patches since this is a major bump and they're no longer
relevant.

Introduce the new RRDTOOL_RRDGRAPH option, in previous versions the
graphing functionality was bundled with the stats/database backend,
however for newer versions this can be disabled.
Since it's normally expected functionality the default is yes, however
all of the buildroot rrd-using applications can work without this.

Signed-off-by: Gustavo Zacarias <gustavo@zacarias.com.ar>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>