package/arptables: add patch to fix musl build issue
Due to clashes in the namespace between the Linux Kernel headers and the
netinet headers building arptables with the musl C library fails.

Best advice from the musl developers is to not include both headers in a
userspace tool (see the thread on the musl mailing list [1], especially [2]).

Since arptables only requires a few definitions from the Linux Kernel headers
we opt-out the clashing header files and define the needed definitions.

Patch sent upstream [3].

Fixes:
http://autobuild.buildroot.net/results/b72/b721e4fa24a0728c20fd2ec0663c2b5981387642/
http://autobuild.buildroot.net/results/4e9/4e9e0ed9454d258ef4333e847dc394ff22378ca4/
http://autobuild.buildroot.net/results/b39/b3951f700ab1664e06e14d95964875fc8e3dd8f4/
http://autobuild.buildroot.net/results/716/716e81f49f8894b7248f394919c8f22efb908277/

[1] http://www.openwall.com/lists/musl/2012/10/09/1
[2] http://www.openwall.com/lists/musl/2012/10/11/1
[3] http://marc.info/?l=netfilter&m=144903021705206&w=2

Cc: Brendan Heading <brendanheading@gmail.com>
Cc: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>