package/pixman: disable tests
Tests are broken on some architectures because of an incomplete fenv.

Since the tests are not really useful in the Buildroot environment, just
disable them.

Fixes a lot of build failures:
    http://autobuild.buildroot.org/results/cd7/cd722620832c28d22eacfc43336f6952b82305ff/
    http://autobuild.buildroot.org/results/3f4/3f4f1e88988d66b052432885b4f044a0eddcee2c/
    ...

Pixman was already autoreconfed (but that was not needed). Since we now
really need to autoreconf it, add a comment stating why we do.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Cc: Bernd Kuhls <bernd.kuhls@t-online.de>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>