ustr: use Debian patches to compile w/ GCC 5.x
To fix issues with compiling with GCC 5.x, switch to utilizing the
Debian patches for cross compiling. Debian has been maintaining
patches for the ustr package since it appears that the ustr project
hasn't been active since 2008.

The Debian patches switch from a makefile based build to autotools
based build. This also means that the patch that was being carried
for cross compiling is no longer needed.

ustr is a dependency of libsemanage, which now builds fine again with
these patches applied.

Fixes:
    http://autobuild.buildroot.org/results/f46/f46430c45973dfe39976453f5a91a555ed7bfa9a/
    http://autobuild.buildroot.org/results/e98/e98c82130efaaed29682cf31c1a754278ff3e547/
    ...and many more

Signed-off-by: Ryan Barnett <ryanbarnett3@gmail.com>
[yann.morin.1998@free.fr: build the shared lib at build time, not
install time]
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>
CC: Arnout Vandecappelle <arnout@mind.be>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>