package/expedite: move to package directory
There is no advantage for efl related packages to share the same version
number anymore (except for Efl and Elementary).

Here are the version number used for the 1.15 stable release:
EFL                     1.15.2
Elementary              1.15.2
Emotion Generic Players 1.15.0
Evas Generic Loaders    1.15.0
Python-EFL              1.15.0

Also, we usually do not have a sub-directory for a family of related
packages which don't share the same version number, so move expedite
to the package directory. Expedite now appears in the
"Graphic libraries and applications (graphic/text)" in the Kconfig menu.

In a followup patch, expedite will be downloaded directly from the 1.15
branch in the git repository since there is no new tarball release after
1.7.0.

Signed-off-by: Romain Naour <romain.naour@openwide.fr>
Reviewed-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>