mysql: don't install libraries in /usr/lib/mysql directory
Installing libraries in a subdirectory of /usr/lib leads to no end of
trouble. It requires either setting a RUN_PATH in the ELF files linked
with it or adding the path to ld.so.conf and calling ldconfig on the
target.

So to simplify things, put everything in /usr/lib instead of
/usr/lib/mysql

[Thomas:
 - tweak commit title.
 - also remove ld.so.conf handling in mysql.mk, since it is no longer
   necessary.]

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>