package/gstreamer1: move comment around, to keep proper indentation
In Kconfig, when there is an item (like a comment) between a symbol and
its dependees, then the indentation is not in effect.

Move the symbol to the top of the file to guarantee the indentation in
menuconfig.

In this case, it means the gstremaer1 plugins are not indented below the
main gstreamer1 prompt.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>