iucode-tool: argp-standalone also needed with musl
The iucode-tool package needs the argp functionality, which is
provided built-in by glibc, but not by uClibc and musl. In uClibc
configurations, it correctly selects the argp-standalone package to
compensate this, but forgets to do so in musl-based
configurations. This commit fixes that by selecting the
argp-standalone package, and adjusting the dependency in the .mk file.

Fixes:

  http://autobuild.buildroot.net/results/7f8626db69500a84a393053a485f04180c565673/

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>